Allowing the container and Entity Manager Factory to be overwritten
In order to use DSM with Symfony I want to use the configuration from
the config directory, rather than the ConfigInterface. To do this I need
to override the EntityManagerFactory, and the Container to include it.

These change should allow me to do this with out breaking anything that
is currently in use