Prevent errors when the collections have not been set
The change tracking policy assumes that the collections have been set,
however when the object is instantiated without using the constructor
this is not guaranteed. When this happens an error will be thrown.

This is a particular problem when using faker to create and hydrate
objects, but could occur in other places

This checks to see if the collection object is null, and if it is
creates a new ArrayCollection before continuing, which removes the
problem