Allowing the repositories to have the correct return types
Should generate these for all of the entities to prevent having to check
for the type each time something is fetched.

Also adding in get and getOneBy commands for when the entity is known to
exist to prevent having to check for null results. This will throw an
exception if nothing is found so use carefully