Allow the Fixtures Loader to be replaced with a new one
At the moment, once a fixture is added once it will get loaded every time
the run method is called. This is less than ideal when the class is
getting passed around a large test suite where different test want
different fixtures.

This adds in a new method that allows the current fixtureLoader to be
replaced with a new empty one.