Adding in a retry connection wrapper
Have got a basic skeleton hard coded into the configuration at the
moment. It wraps each of the connections in a method that checks for
gone away exceptions, and if it finds them tries to restart the
connection before running the query again. If this takes place in a
statement, it tries to rebuild the statement before trying again.

Still to do:

 * Don't hard code this, make it configurable
 * Test that the reconnection actually works
 * Document what this is doing in the README and code