Allow entities to share IDs
At the moment the system is setup to enforce that every ID has to be
unique across all entities.

This will cause problems when we implement inheritance, or deterministic
UUIDs that may require different entities to have the same ID, or use
something other than UUIDs to identify entities.

This starts the process of ensuring that IDs are only unique within each
entity type