Allow non ordered, binary UUIDs
At the moment we are have to use either the binary ordered UUID field or
the non binary UUID field. The binary version requires you to use a UUID
v1, however sometimes you may wish to use a deterministic UUID (i.e. v3
or 5).

This adds the non ordered binary field to Mapping helper and registers
it with doctrine, to allow this to be used.