Provide a standard way to generate a custom number of bespoke fixutes
At the moment the fixture classes are required to generate a fixed
number of entities. It is possible to overwrite the entire class to
generate these, but there are times where you will want to generate a
known set of entities for an individual test.

This adds the functionality to do this to the Abstract fixture loader,
allowing it to be used with all of the fixtures