Make sure logger is present on the connection
It is possible for the Connection to have a logger as well as the Entity
Manager. If it does then currently nothing get logs and the cache
returns an empty object.

This makes sure that the logger is set on both objects and then resets
the loggers to their previous state rather than setting them to null