Only validate single objects, not collections
By default the Valid constraint will call itself on every object in a
collection. This means that each item has to be loaded and itself
validated, which if it the items also contain collections this can
become expensive very quickly when working with large amounts of data.

This changes the default behaviour to not load these up, allowing this
to be overwritten if needed